выделяет валидатор в отдельный модуль. Мелкие правки
