+++ js/main.js
@@ -2,5 +2,7 @@ import { getPhotos } from './photo-description.js';
 import { initGallery } from './gallery.js';
 
 import './form.js';
+import './validation.js';
+
 
 initGallery(getPhotos());
